feat: add `SessionCleanupService` for better session management modularity
