+++ server/src/features/auth/services/user-session/session-cleanup.service.ts
@@ -1,0 +1,111 @@
+import { DbAccess } from "../../../../error";
+import { BaseResult } from "../../../../types";
+import { HashUtil, ResultBuilder } from "../../../../utils";
+import { SessionTokenRepository, UserSessionRepository } from "../repositories";
+
+export class SessionCleanupService {
+  private readonly _sessionTokenRepository: SessionTokenRepository;
+  private readonly _userSessionRepository: UserSessionRepository;
+
+  constructor(
+    sessionTokenRepository: SessionTokenRepository,
+    userSessionTokenRepository: UserSessionRepository
+  ) {
+    this._sessionTokenRepository = sessionTokenRepository;
+    this._userSessionRepository = userSessionTokenRepository;
+  }
+
+  /**
+   * @public
+   * @async
+   * @function tryEndSession
+   * @description Asynchronously attempts to end a session with the hash of a
+   * specified `sessionNumber`.
+   * @param sessionNumber The `sessionNumber` of the session that will be deleted.
+   * @returns A `Promise` resolving to the `id` of the deleted session, or `null` if the
+   * operation fails.
+   */
+  public async endSession(
+    sessionNumber: string
+  ): Promise<
+    | BaseResult.Success<number | undefined, "DB_DELETE">
+    | BaseResult.Fail<DbAccess.ErrorClass>
+  > {
+    try {
+      const deleteResult = await this._userSessionRepository.deleteSessions({
+        scope: "user_session",
+        sessionNumberHash: HashUtil.byCrypto(sessionNumber),
+      });
+
+      const deletedId = deleteResult[0];
+
+      return ResultBuilder.success(deletedId, "DB_DELETE");
+    } catch (err) {
+      const error: DbAccess.ErrorClass = {
+        name: "DB_ACCESS_QUERY_ERROR",
+        message: "Failed deleting session. Please try again later.",
+        cause: err,
+      };
+
+      return ResultBuilder.fail(error);
+    }
+  }
+  /**
+   * @public
+   * @async
+   * @function endIdleSessions
+   * @description Asynchronously ends/deletes idle sessions in the database.
+   * @returns A `Promise` that resolves to an array of `number`s representing
+   * the deleted session ids or `null` if the delete operation fails.
+   */
+  public async endIdleSessions(): Promise<
+    | BaseResult.Success<number[], "DB_DELETE">
+    | BaseResult.Fail<DbAccess.ErrorClass>
+  > {
+    try {
+      const deletedSessionIds =
+        await this._userSessionRepository.deleteSessions({
+          scope: "idle_session",
+        });
+
+      return ResultBuilder.success(deletedSessionIds, "DB_DELETE");
+    } catch (err) {
+      const error: DbAccess.ErrorClass = {
+        name: "DB_ACCESS_QUERY_ERROR",
+        message: "Failed deleting idle sessions.",
+        cause: err,
+      };
+      return ResultBuilder.fail(error);
+    }
+  }
+
+  /**
+   * @public
+   * @async
+   * @function endExpiredSessions
+   * @description Asynchronously ends/deletes all expired sessions in the
+   * database.
+   * @returns A `Promise` that resolves to an array of `number`s representing
+   * the deleted session ids or `null` if the delete operation fails.
+   */
+  public async endExpiredSessions(): Promise<
+    | BaseResult.Success<number[], "DB_DELETE">
+    | BaseResult.Fail<DbAccess.ErrorClass>
+  > {
+    try {
+      const deletedSessionIds =
+        await this._userSessionRepository.deleteSessions({
+          scope: "expired_persistent",
+        });
+
+      return ResultBuilder.success(deletedSessionIds, "DB_DELETE");
+    } catch (err) {
+      const error: DbAccess.ErrorClass = {
+        name: "DB_ACCESS_QUERY_ERROR",
+        message: "Failed deleting idle sessions.",
+        cause: err,
+      };
+      return ResultBuilder.fail(error);
+    }
+  }
+}
